* add priority and multi roles deploy; * refactoring.
